Removed controller logic from category.py and its subclasses.  Categories now return a dict of their heuristic classes as keys and their relative weights as values.  Changed heuristics to use static methods as there was no real reason to create an instance.  All evaluation logic is now in common/analysis.py
