+++ api/app.py
@@ -5,7 +5,6 @@ from flask import Flask, Blueprint, g
 from flask_cors import CORS
 from flask_restful import Resource, Api, url_for
 from flask_pymongo import PyMongo
-from common import auth
 
 from controllers import input, status, results
