+++ api/models/heuristics/documentlength.py
@@ -4,7 +4,8 @@ from readcalc import readcalc
 # Formal 4
 # Ensure reasonable document length for target user
 class DocumentLength(Heuristic):
-    def score(self, eula):
+    @staticmethod
+    def score(eula):
         text = eula.text
         calc = readcalc.ReadCalc(text)
         words = len(calc.get_words())
