+++ api/models/heuristics/easeofnavigation.py
@@ -3,5 +3,6 @@ from models.heuristic import Heuristic
 # Formal 3
 # Ensure ease of user navigation.
 class EaseOfNavigation(Heuristic):
-    def score(self, eula):
+    @staticmethod
+    def score(eula):
         return {'score': -1, 'max': 4, 'reason': 'Not implemented'}
