+++ api/models/heuristics/mobileaccessibility.py
@@ -3,5 +3,6 @@ from models.heuristic import Heuristic
 # Procedural 1a
 # Ensure contrast of EULA link with background on mobile devices
 class MobileAccessibility(Heuristic):
-    def score(self, eula):
+    @staticmethod
+    def score(eula):
         return {'score': -1, 'max': 4, 'reason': 'Not implemented'}
